Fix OperationIdentifier hash to use uint32_t
The hash function for OperationIdentifier was incorrectly hashing k.id
as uint16_t when id is actually uint32_t, causing truncation and hash
collisions for operation IDs exceeding 65,535.

This would cause issues in any hash maps or sets using OperationIdentifier
as a key in the SSA phase, leading to incorrect lookups and transformations.

+++ nautilus/src/nautilus/compiler/ir/operations/Operation.hpp
@@ -133,7 +133,7 @@ struct hash<nautilus::compiler::ir::OperationIdentifier> {
 	std::size_t operator()(const nautilus::compiler::ir::OperationIdentifier& k) const {
 		using std::hash;
 		using std::size_t;
-		return hash<uint16_t>()(k.id);
+		return hash<uint32_t>()(k.id);
 	}
 };
 } // namespace std
